# openapi.yaml
openapi: 3.0.3
info:
  title: GoalSeek Personal Assistant API
  description: API for personal AI assistant that helps users achieve their goals and optimize daily performance
  version: 1.0.0
  contact:
    name: GoalSeek Team
    url: https://goalseek.jmoore.info
servers:
  - url: https://goalseek.jmoore.info/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
components:
  schemas:
...
    Goal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        progress:
          type: number
          minimum: 0
          maximum: 100
        deadline:
          type: string
          format: date
        category:
          type: string
          enum: [Learning, Fitness, Career, Finance, Relationships, Health]
        priority:
          type: string
          enum: [High, Medium, Low]
        type:
          type: string
          enum: [short-term, long-term]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - title
        - progress
        - category
        - priority
        - type
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        completed:
          type: boolean
          default: false
        priority:
          type: string
          enum: [High, Medium, Low]
        category:
          type: string
        estimated_time:
          type: number
          description: Estimated time in minutes
        deadline:
          type: string
          format: date-time
        scheduled_time:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - title
        - completed
        - priority
        - category
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /goals:
    get:
      summary: Get user goals
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [short-term, long-term]
          description: Filter by goal type
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
      responses:
        "200":
          description: Goals retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"

    post:
      summary: Create a new goal
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Goal"
      responses:
        "201":
          description: Goal created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{goalId}:
    get:
      summary: Get a specific goal
      security:
        - BearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Goal retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

    put:
      summary: Update a goal
      security:
        - BearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Goal"
      responses:
        "200":
          description: Goal updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"

    delete:
      summary: Delete a goal
      security:
        - BearerAuth: []
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Goal deleted successfully
  /tasks:
    get:
      summary: Get user tasks
      security:
        - BearerAuth: []
      parameters:
        - name: completed
          in: query
          schema:
            type: boolean
          description: Filter by completion status
        - name: priority
          in: query
          schema:
            type: string
            enum: [High, Medium, Low]
          description: Filter by priority
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter by date
      responses:
        "200":
          description: Tasks retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

    post:
      summary: Create a new task
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        "201":
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      summary: Get a specific task
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    put:
      summary: Update a task
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Task"
      responses:
        "200":
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

    delete:
      summary: Delete a task
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Task deleted successfully
...